Work around old SELinux / AppArmor bug
Linux thinks /dev/zero is a real file with hostile contents, so both
AppArmor and SELinux treat it as such when enforcing their policies (of
questionable merit). This is simply an old bug in the Linux kernel. As
far as POSIX is concerned, the portable method for creating an anonymous
mapping is to map /dev/zero. With this Linux kernel bug, the portable
approach doesn't work. So instead use the non-portable MAP_ANONYMOUS
flag.

I've run into this issue a few times before:

A JIT Compiler Skirmish with SELinux
https://nullprogram.com/blog/2018/11/15/